+++ simplescn/pwrequester/cmd.py
@@ -1,0 +1,7 @@
+#! /usr/bin/env python3
+
+from getpass import getpass
+
+def pwcallmethod(msg):
+    return getpass(msg+":\n")
+