+++ simplescn/tools/__init__.py
@@ -31,9 +31,10 @@ except ImportError:
     psutil = None
 
 
-from .. import config, pwcallmethod
+from .. import config
+from ..pwrequester import pwcallmethod
 from ..config import isself
-from .. import AddressError, AddressEmptyError, AddressLengthError, EnforcedPortError
+from ..exceptions import AddressError, AddressEmptyError, AddressLengthError, EnforcedPortError
 ##### init ######
 
 def generate_certs(_path):
@@ -591,11 +592,10 @@ def safe_mdecode(inp, encoding, charset="utf-8"):
         logging.error(exc)
         return None
 
-@functools.lru_cache(maxsize=256)
-def genc_error(err: str):
-    """ Quick error, be careful: if used for more different strings than maxsize, old entries are deleted """
-    #assert isinstance(err, str), "genc_error only accepts str"
-    return generate_error(err, False)
+def quick_error(err: str):
+    """ Quick error """
+    #assert isinstance(err, str), "quick_error only accepts str"
+    return {"msg": err, "type": ""}
 
 def generate_error(err, withstack=True):
     """ generate error from string/exception/... """